Refactored few spots in the GUI to remove the usage of singletons
The previous code relied on the static attribute `Store::instance` to be
populated. If this attribute wasn't populated in time, the application
falls apart.

To combat this, a new method `makeLoader(): FXMLLoader` was implemented,
which uses a custom controller factory, allowing for parameters to be
passed to the controller upon construction.