Began integrating new network module into application
As of this commit, connections to a server are now initiated
using clients created by `ClientSelector::connect`.

Additionally, two new actions were introduced. The first one being
`LoggedInAction`, which is issued by the client itself, when it
successfully logged onto the server.

The second one being `ApplicationExitAction`, which is issued, when
the application closes. The reducer is supposed to stop threads and
free up any resources.

This commit/branch addresses #4.