Fixed bug in IOHandler.java
The IOHandler assumed, that each command was received individually. On faster machines running the server locally this is not the case.

This resulted the deserializer in receiving strings such as "command1\ncommand2", which it was not intented for.

The IOHandler not splits strings into commands (i.e. lines of commands) and feeds them one by one into the deserializer.